Base.run

Replication_Monge_et_al_2019.create_figure_1

Replication_Monge_et_al_2019.create_figure_2

Replication_Monge_et_al_2019.create_figure_4

Replication_Monge_et_al_2019.create_table_1

Replication_Monge_et_al_2019.create_table_3

Replication_Monge_et_al_2019.create_table_4

Replication_Monge_et_al_2019.create_table_5

Replication_Monge_et_al_2019.delete_all

Replication_Monge_et_al_2019.delete_figure_1

Replication_Monge_et_al_2019.delete_figure_2

Replication_Monge_et_al_2019.delete_figure_4

Replication_Monge_et_al_2019.delete_table_1

Replication_Monge_et_al_2019.delete_table_3

Replication_Monge_et_al_2019.delete_table_4

Replication_Monge_et_al_2019.delete_table_5

Replication_Monge_et_al_2019.test_function

Base.run — Method

The function run() call all the results obtained from the replication attempt. It produces files and save them in the output folder.

Replication_Monge_et_al_2019.create_figure_1 — Method

The function create_figure_1() creates a png containing the replication result of the Figure 1.

The png file is created within an output folder.

Replication_Monge_et_al_2019.create_figure_2 — Method

The function create_figure_2() creates a png containing the replication result of the Figure 2.

The png file is created within an output folder.

Replication_Monge_et_al_2019.create_figure_4 — Method

The function create_figure_4() creates a png containing the replication result of the Figure 4.

The png file is created within an output folder.

Replication_Monge_et_al_2019.create_table_1 — Method

The function create_table_1() creates a pdf containing the replication result of the Figure 1.

The pdf file is created within an output folder.

Replication_Monge_et_al_2019.create_table_3 — Method

The function create_table_3() creates a pdf containing the replication result of the Figure 3.

The pdf file is created within an output folder.

Replication_Monge_et_al_2019.create_table_4 — Method

The function create_table_4() creates a pdf containing the replication result of the Figure 3.

The pdf file is created within an output folder.

Replication_Monge_et_al_2019.create_table_5 — Method

The function create_table_5() creates a pdf containing the replication result of the Figure 5.

The pdf file is created within an output folder.

Replication_Monge_et_al_2019.delete_all — Method

The function delete_all() deletes the all the files containing the replication results of the package, if the present working directory has a folder output containing it.

Warning : The function throws an error if any output file is missing.

Replication_Monge_et_al_2019.delete_figure_1 — Method

The function delete_figure_1() deletes the pdf file containing the replication result of Figure 1, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_figure_2 — Method

The function delete_figure_2() deletes the png file containing the replication result of Figure 2, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_figure_4 — Method

The function delete_figure_4() deletes the png file containing the replication result of Figure 4, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_table_1 — Method

The function delete_table_1() deletes the pdf file containing the replication result of Figure 1, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_table_3 — Method

The function delete_table_3() deletes the pdf file and oher building blocks containing the replication result of Table 3, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_table_4 — Method

The function delete_table_4() deletes the pdf file and oher building blocks containing the replication result of Table 4, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.delete_table_5 — Method

The function delete_table_5() deletes the pdf file and oher building blocks containing the replication result of Table 5, if the present working directory has a folder output containing it.

Replication_Monge_et_al_2019.test_function — Method

This function is a test. We can use it as following :

test_function()

It prints "The ReplicationMongeetal2019 module is loaded.".